fix: moved components to match Docusaurus conventions
