+++ website/src/pages/playground/index.js
@@ -9,11 +9,10 @@ import React, { useEffect, useState, useReducer } from 'react';
 import classnames from 'classnames';
 import MonacoEditor from 'react-monaco-editor';
 import Layout from '@theme/Layout';
-import useTheme from './useTheme';
-import Worker from 'worker-loader!./worker.js';
+import Worker from 'worker-loader!@site/src/workers/hermes.js';
+import useTheme from '@site/src/hooks/useTheme';
+import useWindowSize from '@site/src/hooks/useWindowSize';
 import styles from './styles.module.css';
-import RunIcon from './runIcon';
-import useWindowSize from './useWindowSize';
 
 const vsDarkTheme = {
   base: 'vs-dark',
@@ -30,11 +29,6 @@ function reducer(state, action) {
   switch (action.type) {
     case 'request':
       return { ...state, loading: true, lastTime: new Date() };
-    case 'log':
-      return {
-        ...state,
-        output: state.output + action.log,
-      };
     case 'success':
       return {
         ...state,
@@ -58,7 +52,7 @@ function Playground() {
     initialState
   );
   const [input, setInput] = useState('const a = 1; \nprint(a);');
-  const [args, setArgs] = useState('-help');
+  const [args, setArgs] = useState('-O -dump-bytecode');
   const windowSize = useWindowSize();
   const theme = useTheme();
 
@@ -91,9 +85,6 @@ function Playground() {
   useEffect(() => {
     worker.onmessage = function(e) {
       switch (e.data[0]) {
-        case 'log':
-          dispatch({ type: 'log', log: e.data[1].log });
-          break;
         case 'result':
           dispatch({
             type: 'success',
@@ -102,9 +93,11 @@ function Playground() {
           break;
       }
     };
+
+    run(args);
   }, []);
 
-  function run() {
+  function run(args) {
     if (loading) {
       return;
     }
@@ -116,11 +109,19 @@ function Playground() {
 
   function handleSubmit(e) {
     e.preventDefault();
-    run();
+    run(args);
   }
 
   function handleArgsChange(evt) {
     setArgs(evt.target.value);
+  }
+
+  function handleClickRun() {
+    run(args);
+  }
+
+  function handleClickHelp() {
+    run('-help');
   }
 
   function onEditorWillMount(monaco) {
@@ -129,25 +130,39 @@ function Playground() {
 
   return (
     <Layout title="Hermes" description="Hermes Playground" noFooter={true}>
-      <form
-        onSubmit={handleSubmit}
-        className={classnames(styles.argsInputContainer)}
-      >
-        <input
-          className={classnames(styles.argsInputField)}
-          name="args"
-          placeholder="args"
-          type="text"
-          value={args}
-          onChange={handleArgsChange}
-        />
-        <button onClick={run} className={classnames(styles.argsIconContainer)}>
-          <RunIcon loading={loading} />
-        </button>
-        <span className={classnames(styles.ellapsed)}>{ellapsed}</span>
-      </form>
+      <div className={classnames(styles.headerContainer)}>
+        <form
+          onSubmit={handleSubmit}
+          className={classnames(styles.argsInputContainer)}
+        >
+          <input
+            className={classnames(styles.argsInputField)}
+            name="args"
+            placeholder="args"
+            type="text"
+            value={args}
+            onChange={handleArgsChange}
+          />
+          <button
+            onClick={handleClickRun}
+            className={classnames(styles.argsIconContainer)}
+          >
+            <RunIcon loading={loading} />
+          </button>
+        </form>
 
-      <div className={styles.row}>
+        <div>
+          <button
+            onClick={handleClickHelp}
+            className={classnames(styles.helperButton)}
+          >
+            ?
+          </button>
+          <span>{ellapsed}</span>
+        </div>
+      </div>
+
+      <div className={classnames(styles.row)}>
         <MonacoEditor
           {...(windowSize.width > 600 ? editorLayout.lg : editorLayout.xs)}
           language="javascript"
@@ -171,6 +186,41 @@ function Playground() {
         />
       </div>
     </Layout>
+  );
+}
+
+function RunIcon({ loading }) {
+  if (loading) {
+    return (
+      <i aria-label="icon: load">
+        <svg
+          viewBox="0 0 1024 1024"
+          data-icon="loading"
+          width="1.5em"
+          height="1.5em"
+          fill="currentColor"
+          aria-hidden="true"
+          className={classnames(styles.spinner)}
+        >
+          <path d="M988 548c-19.9 0-36-16.1-36-36 0-59.4-11.6-117-34.6-171.3a440.45 440.45 0 0 0-94.3-139.9 437.71 437.71 0 0 0-139.9-94.3C629 83.6 571.4 72 512 72c-19.9 0-36-16.1-36-36s16.1-36 36-36c69.1 0 136.2 13.5 199.3 40.3C772.3 66 827 103 874 150c47 47 83.9 101.8 109.7 162.7 26.7 63.1 40.2 130.2 40.2 199.3.1 19.9-16 36-35.9 36z"></path>
+        </svg>
+      </i>
+    );
+  }
+
+  return (
+    <i aria-label="icon: run">
+      <svg
+        viewBox="0 0 1200 1200"
+        data-icon="run"
+        width="1.5em"
+        height="1.5em"
+        fill="currentColor"
+        aria-hidden="true"
+      >
+        <path d="M 600,1200 C 268.65,1200 0,931.35 0,600 0,268.65 268.65,0 600,0 c 331.35,0 600,268.65 600,600 0,331.35 -268.65,600 -600,600 z M 450,300.45 450,899.55 900,600 450,300.45 z" />
+      </svg>
+    </i>
   );
 }
